Turn addresses from u32s into u16s as they should be...

+++ src/cpu/instructions/accumulator.rs
@@ -16,27 +16,27 @@ pub trait Accumulator {
 
 impl Accumulator for Processor {
     fn lda_indirect_x(&mut self, memory: &Memory) -> () {
-        let indirect_addr: u32 = self.addr_indirect_x(memory);
+        let indirect_addr: u16 = self.addr_indirect_x(memory);
         let byte_value = self.read_byte(memory, indirect_addr);
 
         self.set_register(Accumulator, byte_value)
     }
 
     fn lda_indirect_y(&mut self, memory: &Memory) -> () {
-        let indirect_addr: u32 = self.addr_indirect_y(memory);
-        let byte_value = self.read_byte(memory, indirect_addr as u32);
+        let indirect_addr: u16 = self.addr_indirect_y(memory);
+        let byte_value = self.read_byte(memory, indirect_addr);
 
         self.set_register(Accumulator, byte_value);
     }
 
     fn sta_indirect_x(&mut self, memory: &mut Memory) -> () {
-        let indirect_addr: u32 = self.addr_indirect_x(memory);
+        let indirect_addr: u16 = self.addr_indirect_x(memory);
 
         memory.write_byte(self.accumulator, indirect_addr, &mut self.cycles);
     }
 
     fn sta_indirect_y(&mut self, memory: &mut Memory) -> () {
-        let indirect_addr: u32 = self.addr_indirect_y(memory);
+        let indirect_addr: u16 = self.addr_indirect_y(memory);
 
         memory.write_byte(self.accumulator, indirect_addr, &mut self.cycles);
     }
